not castin to int

+++ core/src/com/mygdx/game/singelton/Settings.java
@@ -15,7 +15,7 @@ public final class Settings {
     private Queue<Integer> nextObstacles;
     private Queue<Integer> nextObstacleHeight;
     private Queue<Integer> nextTime;
-    private int obstacleTime;
+    private long obstacleTime;
 
     public static Settings getInstance(){
         return INSTANCE;
@@ -59,7 +59,7 @@ public final class Settings {
     public void multiplayerReady(String id){
         this.id = id;
         this.mutiplayerReady = true;
-        this.obstacleTime = (int)System.currentTimeMillis();
+        this.obstacleTime = System.currentTimeMillis();
         System.out.println("Multiplayer is set to ready");
     }
 
@@ -130,11 +130,11 @@ public final class Settings {
         return nextTime.size;
     }
 
-    public void setObstacleTime(int time){
+    public void setObstacleTime(long time){
         obstacleTime = time;
     }
 
-    public int getObstacleTime(){
+    public long getObstacleTime(){
         return obstacleTime;
     }
